tabCompleterをCommand処理から独立したclassに切り分け
